version 2.2 eco system, coin icons, UI improvement

+++ app/src/main/java/model/Nickels.java
@@ -29,6 +29,6 @@ public class Nickels {
     }
 
     void SetAmount(int inputVal) {
-        amount += (int) (inputVal / val);
+        amount += inputVal / val;
     }
 }
